refactor(cli): extract daemon and headless modules from main.rs
Move daemon lifecycle (spawn, start, stop, status) to commands/daemon.rs
and headless/snapshot-tui modes to commands/headless.rs. Reduces main.rs
from 1006 to 540 lines.
