+++ crates/astrid-cli/src/commands/mod.rs
@@ -3,7 +3,9 @@
 pub(crate) mod capsule;
 pub(crate) mod chat;
 pub(crate) mod config;
+pub(crate) mod daemon;
 pub(crate) mod distro;
+pub(crate) mod headless;
 pub(crate) mod init;
 pub(crate) mod self_update;
 pub(crate) mod sessions;
